Update scheduling to new schema. Reduce amount of queued jobs.
